% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{find_default_app}
\alias{find_default_app}
\title{Find default shiny app}
\usage{
find_default_app()
}
\description{
Wama uses the first non-empty return of the below as your default shiny app:
}
\details{
\enumerate{
\item The \code{Config/wama/defaultShinyApp} field of the \code{DESCRIPTION} in the working directory.

To set this for your package, add this to your \code{DESCRIPTION}:\preformatted{Config/wama/defaultShinyApp: myApp()
}
\item The \code{wama.default.shiny.app} option.

This is best used interactively (\code{options(wama.default.app = myApp())})
or set in your \code{.Rprofile}.
}
}
