% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wama-package.R
\docType{package}
\name{wama-package}
\alias{wama}
\alias{wama-package}
\title{wama: Automating loading workflows during local development}
\description{
Provides helpers to speed up iterations during local development.
    Whenever relevant files are changed on disc, a loading workflow is triggered.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.maxheld.de/wama}
  \item \url{https://github.com/maxheld83/wama}
  \item Report bugs at \url{https://github.com/maxheld83/wama/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{held@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

}
\keyword{internal}
