% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wama-package.R
\docType{package}
\name{wama-package}
\alias{wama}
\alias{wama-package}
\title{Automating loading workflows during interactive development}
\description{
Provides helpers to speed up iterations during local development.
    Whenever relevant files are changed on disc, a loading workflow is triggered.
}
\section{Package options}{


Wama uses the following following \code{\link[=options]{options()}} to configure behaviour:
\itemize{
\item \code{wama.default.app}:
The default shiny application to run in the development context
of some package.

If your package includes only \emph{one} shiny application,
you can conveniently set declare this to be used in your \code{zzz.R}
(the filename is a convention):\if{html}{\out{<div class="r">}}\preformatted{.onLoad <- function(libname, pkgname) options(wama.default.app = myapp())
}\if{html}{\out{</div>}}

When your package is loaded \emph{last} (as is typically the case in development),
all wama functions will default to this default app of your package.

If you have more than one app, you can always change the default app
interactively using \code{options(wama.default.app = some_app())}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.maxheld.de/wama}
  \item \url{https://github.com/maxheld83/wama}
  \item Report bugs at \url{https://github.com/maxheld83/wama/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{held@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

}
